XP-1783 Multiple duplicated requests when selecting content in Content Manager
