+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/content-manager/js/app/view/detail/_module.ts
@@ -1,4 +1,8 @@
+///<reference path='button/_module.ts' />
+///<reference path='NonMobileDetailsPanelsManager.ts' />
+///<reference path='WidgetsSelectionRow.ts' />
 ///<reference path='DetailsPanel.ts' />
+///<reference path='ActiveDetailsPanelManager.ts' />
 ///<reference path='WidgetView.ts' />
 ///<reference path='WidgetItemView.ts' />
 ///<reference path='StatusWidgetItemView.ts' />
