+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/content-manager/js/app/view/detail/button/_module.ts
@@ -1,0 +1,2 @@
+///<reference path='MobileDetailsPanelToggleButton.ts' />
+///<reference path='InfoWidgetToggleButton.ts' />
