+++ modules/admin/admin-ui/src/main/resources/web/admin/common/styles/apps/content/view/details/details.less
@@ -417,7 +417,7 @@
       }
 
       .widgets-selection-row .widget-selector {
-        width: calc(~'100% - 40px');
+        width: calc(~'100% - 42px');
       }
     }
   }
