+++ modules/admin/admin-ui/src/main/resources/web/admin/common/styles/apps/content/view/mobile-content-item-statistics-panel.less
@@ -105,7 +105,7 @@
       }
 
       .widget-selector {
-        width: calc(~'100% - 40px');
+        width: calc(~'100% - 42px');
 
         .viewer {
           font-size: 12px;
